Move the RAG augmentor bridge into the langchain4j-embeddingstore extension
The bridge producing RetrievalAugmentor beans and binding @EmbeddingStoreName CDI
beans into the Camel registry lived in camel-quarkus-support-langchain4j, so it was
active for all six extensions depending on that artifact. Neither half has a consumer
in the other five: embeddingStore is a URI option on camel-langchain4j-embeddingstore
alone, and no Camel component consumes a RetrievalAugmentor. It is now registered only
when camel-quarkus-langchain4j-embeddingstore is a dependency.

* Move RagRetrievalFilterSupplier, RagAugmentorName, RagBridgeConfig and
  DefaultRetrievalAugmentorSupplier to the extension runtime, and split the RAG half
  of QuarkusLangchain4jRecorder into Langchain4jEmbeddingstoreRecorder
* Move the six RAG build steps, BeanCensus and AugmentorDefinition into
  Langchain4jEmbeddingstoreRagProcessor, along with EasyRagPresent and its unit test.
  SupportQuarkusLangchain4jProcessor drops from 435 to 152 lines
* Document the narrowed activation and the RagRetrievalFilterSupplier package change
  in the 3.40.0 migration guide
