+++ docs/modules/ROOT/pages/reference/extensions/langchain4j-embeddingstore.adoc
@@ -165,3 +165,39 @@ As such, one might need to import the `langchain4j-bom` before other related Cam
 
 WARNING: At present, this extension is neither tested with nor intended to be used in conjunction with any Quarkus LangChain4j extensions.
 Consequently, both JVM and native modes may exhibit unexpected behaviour or fail to function correctly in such configurations.
+
+[id="extensions-langchain4j-embeddingstore-additional-camel-quarkus-configuration"]
+== Additional Camel Quarkus configuration
+
+[width="100%",cols="80,5,15",options="header"]
+|===
+| Configuration property | Type | Default
+
+
+a|icon:lock[title=Fixed at build time] [[quarkus-camel-langchain4j-rag-augmentors-augmentors-embedding-store-name]]`link:#quarkus-camel-langchain4j-rag-augmentors-augmentors-embedding-store-name[quarkus.camel.langchain4j.rag.augmentors."augmentors".embedding-store-name]`
+
+CDI bean name of the `EmbeddingStore` to use.
+Matches beans annotated with `@Named("name")` or `@EmbeddingStoreName("name")`.
+| `string`
+| required icon:exclamation-circle[title=Configuration property is required]
+
+a|icon:lock[title=Fixed at build time] [[quarkus-camel-langchain4j-rag-augmentors-augmentors-embedding-model-name]]`link:#quarkus-camel-langchain4j-rag-augmentors-augmentors-embedding-model-name[quarkus.camel.langchain4j.rag.augmentors."augmentors".embedding-model-name]`
+
+CDI bean name of the `EmbeddingModel` to use.
+When not set, the default (unnamed) `EmbeddingModel` is used.
+| `string`
+| 
+
+a|icon:lock[title=Fixed at build time] [[quarkus-camel-langchain4j-rag-augmentors-augmentors-default]]`link:#quarkus-camel-langchain4j-rag-augmentors-augmentors-default[quarkus.camel.langchain4j.rag.augmentors."augmentors".default]`
+
+Marks this augmentor as the one `@RegisterAiService` AI services use when they
+do not select an augmentor explicitly. Exactly one augmentor must be marked when more
+than one is configured — otherwise the build fails, because an unmarked ambiguity
+would silently disable RAG for every AI service in the application.
+| `boolean`
+| `false`
+|===
+
+[.configuration-legend]
+{doc-link-icon-lock}[title=Fixed at build time] Configuration property fixed at build time. All other configuration properties are overridable at runtime.
+
