+++ extensions-support/langchain4j/runtime/src/main/java/org/apache/camel/quarkus/component/support/langchain4j/QuarkusLangchain4jRecorder.java
@@ -16,69 +16,15 @@
  */
 package org.apache.camel.quarkus.component.support.langchain4j;
 
-import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationTargetException;
-import java.util.List;
-import java.util.function.Supplier;
 
-import dev.langchain4j.data.segment.TextSegment;
 import dev.langchain4j.guardrail.Guardrail;
-import dev.langchain4j.rag.RetrievalAugmentor;
-import dev.langchain4j.store.embedding.EmbeddingStore;
-import io.quarkus.arc.Arc;
-import io.quarkus.arc.InstanceHandle;
 import io.quarkus.runtime.RuntimeValue;
 import io.quarkus.runtime.annotations.Recorder;
-import jakarta.enterprise.inject.Any;
-import jakarta.enterprise.inject.spi.Bean;
-import jakarta.enterprise.util.TypeLiteral;
-import org.apache.camel.spi.Registry;
 import org.jboss.logging.Logger;
 
 @Recorder
 public class QuarkusLangchain4jRecorder {
-
-    private static final Logger LOG = Logger.getLogger(QuarkusLangchain4jRecorder.class);
-    private static final String EMBEDDING_STORE_NAME_CLASS = "io.quarkiverse.langchain4j.EmbeddingStoreName";
-    @SuppressWarnings("serial")
-    private static final TypeLiteral<EmbeddingStore<TextSegment>> EMBEDDING_STORE_TYPE = new TypeLiteral<>() {
-    };
-
-    public void registerNamedEmbeddingStores(RuntimeValue<Registry> camelRegistry) {
-        List<InstanceHandle<EmbeddingStore<TextSegment>>> handles = Arc.container()
-                .listAll(EMBEDDING_STORE_TYPE, Any.Literal.INSTANCE);
-
-        Registry registry = camelRegistry.getValue();
-        for (InstanceHandle<EmbeddingStore<TextSegment>> handle : handles) {
-            Bean<?> bean = handle.getBean();
-            if (bean == null) {
-                continue;
-            }
-            for (Annotation qualifier : bean.getQualifiers()) {
-                if (EMBEDDING_STORE_NAME_CLASS.equals(qualifier.annotationType().getName())) {
-                    try {
-                        String name = (String) qualifier.annotationType()
-                                .getMethod("value").invoke(qualifier);
-                        if (name != null && !name.isEmpty()) {
-                            // The identity check keeps a bean carrying both @Named("x") and
-                            // @EmbeddingStoreName("x") from triggering a spurious warning
-                            EmbeddingStore<?> existing = registry.lookupByNameAndType(name, EmbeddingStore.class);
-                            if (existing != null && existing != handle.get()) {
-                                LOG.warnf(
-                                        "The Camel registry already resolves a different EmbeddingStore under the name \"%s\"; lookups by that name may not return the @EmbeddingStoreName(\"%s\") bean",
-                                        name, name);
-                            }
-                            // Bind via supplier so a store never referenced by a route is not instantiated
-                            registry.bind(name, EmbeddingStore.class, handle::get);
-                            LOG.debugf("Registered @EmbeddingStoreName(\"%s\") in Camel registry", name);
-                        }
-                    } catch (ReflectiveOperationException e) {
-                        LOG.warnf(e, "Failed to extract name from @EmbeddingStoreName qualifier");
-                    }
-                }
-            }
-        }
-    }
 
     public RuntimeValue<Guardrail<?, ?>> instantiateGuardrails(Class<Guardrail<?, ?>> guardrailClass) {
         try {
@@ -89,11 +35,5 @@ public class QuarkusLangchain4jRecorder {
                     "Can not instantiate guardrail of class %s", guardrailClass.getName());
             return null;
         }
-    }
-
-    /** @param augmentorName {@code null} for the auto-produced default augmentor. */
-    public Supplier<RetrievalAugmentor> createRetrievalAugmentorSupplier(
-            String embeddingStoreName, String embeddingModelName, String augmentorName) {
-        return new DefaultRetrievalAugmentorSupplier(embeddingStoreName, embeddingModelName, augmentorName);
     }
 }
