+++ extensions/langchain4j-embeddingstore/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -29,6 +29,8 @@ metadata:
   guide: "https://camel.apache.org/camel-quarkus/latest/reference/extensions/langchain4j-embeddingstore.html"
   categories:
   - "integration"
+  config:
+  - "quarkus.camel.langchain4j.rag"
   status: "preview"
   integrates:
     - name: "Camel"
