+++ integration-tests/langchain4j-rag-bridge-ql4j/src/test/java/org/apache/camel/quarkus/component/langchain4j/ragbridge/it/MultiAugmentorProfile.java
@@ -24,7 +24,7 @@ public class MultiAugmentorProfile implements QuarkusTestProfile {
     @Override
     public Map<String, String> getConfigOverrides() {
         // Two augmentors require exactly one designated default — without the marking the build
-        // fails (see SupportQuarkusLangchain4jProcessor.resolveDesignatedDefault and its test)
+        // fails (see Langchain4jEmbeddingstoreRagProcessor.resolveDesignatedDefault and its test)
         return Map.of(
                 "quarkus.camel.langchain4j.rag.augmentors.products.embedding-store-name", "products",
                 "quarkus.camel.langchain4j.rag.augmentors.products.default", "true",
